Fix wrapAround with adaptiveHeight; move hooks; remove eslint-disable
